Move Sink to actions.py
